+++ arches_her/data_access/common.py
@@ -78,14 +78,9 @@ def serialize(obj):
     return obj
 
 
-class CustomJSONEncoder(json.JSONEncoder):
-    def default(self, obj):
-        return serialize(obj)
-
-
 def generate_json(data: str) -> str:
     processed_data = serialize(data)
-    return json.dumps(processed_data, cls=CustomJSONEncoder, indent=4)
+    return processed_data
 
 
 def call_hapi_get_descriptions(resource_instance_id: uuid.UUID) -> Optional[List[Description]]:
